Generic retry logic to all Matrix requests
This makes it possible for the HS to go down and everything will
continue where the were left off.

All IRC events will be processed and flushed to HS when it comes
back up.

If this causes issues it may need to be made per-event so some
requests fail fast rather than in 30 minutes.