Add delay to ident response to hack around racing
The way asyncio works by default makes it pretty hard to avoid the
server sending an ident request before the connection is fully
established.

Small delay before responding will "fix" it as it gives asyncio
time to finish it up.

If this doesn't work in the long run will find a better solution.

Fixes #79