Implement IRC->Matrix and IRC->IRC pills
This is really the best effort we can do. IRC nicks are used to do
the mapping of pills.

If a Matrix user has an irc nick of "foo" and bar says "hey, foo!"
it will replace "foo" with a pill having the displayname of the
Matrix user.

IRC->IRC works with nicks only.

Fixes #28