Fix lazy member joining racing
By pure chance the displayname update feature kicked in when lazy
joining was enabled and it was _mostly_ run before the puppet was
lazily joined to the room.

This had a few side effects:

 - IRC puppets were always registered when they joined a channel
 - the code to lazily join puppets mistakenly did not ensure
   the IRC puppets are registered as it worked by luck without

Now the displayname cache is ignored when a puppet is in lazy
members list and it will start being updated only after joining.