Invalidate PM rooms when you leave them
This was intentionally enabled before but I believe it's better for
preserving privacy (as much as you can with IRC) that when you
close a PM the room will be garbage collected.

Additionally with autoquery setting disable this would make it
impossible to prevent someone on IRC from reinviting you an old PM.

When starting up the bridge after this change it will likely throw
away some rooms on startup.