Implement rejoin on invite and kick
Rejoin existing rooms on invite or kick. Invite is enabled by
default but kick is not as it is likely not desired for anything
else than plumbs.

Join on invite can be used for "knocking" on rooms with AUTOCMD
on connect to receive an invite. It's also useful if you get
accidentally kicked from a channel and are invited back if you
don't leave the Matrix room.