Clear per-channel send queue when kicked or parted
If you in bouncer mode or the relaybot gets kicked for flooding it
is wise to clear up the send queue for the particular channel to
prevent sending events that either wont reach the channel or would
continue the flood after a rejoin (which may be automatic).

A new STOP command for channels has been added for a bouncer mode
user to clear the queue manually in the case of accidental long
message.