Always ensure talking IRC users join the Matrix room
Make lazy members list a permanent one keeping track of all users
that should be in the room. This works as a failsafe to rejoin
Matrix users that are missing from the room for any reason.

For plumb moderation this means kicking IRC ghosts without ops on
IRC will cause them to rejoin the room when they talk preventing
the fallback text from being used for more consistent behavior.

Fixes #176