Support specifying TLS cipher list per server
Some servers with weird TLS configurations don't accept any of the ciphers
in the default Python ssl module cipher list. To connect to such a server,
it is necessary to specify a custom cipher list, so add an option to the
ADDSERVER command to specify a per-server custom cipher list and pass it
into the SSL context object if set.