Send multiple JOIN commands if 512 bytes message length is violated
As messages are limited to 512 bytes including CR/LF, the JOIN
command fails if the amount of channels intended to be joined is too
large.

This commit allows joining a larger number of rooms by splitting the
channels to be joined among multiple JOIN commands instead of
running only one join commands which violates the message length.